Update editorial-outline.md
This is excellent, great to see this proposal becoming more concrete. I made the mention of licenses a bit more forceful, since that's important for enabling sharing and solving the problem of troubleshooting for others, and added a mention of version control, added some references (no doubt lots more will be needed). Fixed a few typos