Update editorial-outline.md
A note on etiquette, this might not be be the right section, but I think it's important topic to raise in order to change the norms. Many other published comments on reproducibility gloss over the social mechanics of sharing code and data, which leaves researchers with a lot of anxiety about how it's all actually going to work in the fog of day-to-day practice.