Remove section on "current position" describing review
process.

This will be described separately by the journal, once they have
decided what to do.  Text deleted:

\subsection*{Current position}

To help the editors of \textit{Nature Neuroscience} and
\textit{Scientific Data} to evaluate if appropriate code is being
shared, we are now helping to run a one-year trial that adds a small
extra step to the manuscript review process.  During our \textit{code
  availability review}, a computational expert will be asked to check
what code has been provided, and to write a brief report commenting on
its quality according to criteria outlined below (``Simple steps to
help you share your code'').  The report will be sent to the editor
and returned to the authors along with the referees' reviews. There is
no expectation that the expert will re-run the programs, as often it
can take days merely to recreate the computational environment, and
computations could take many months to complete.  Furthermore, this
report will not prevent publication of a manuscript, as the current
policy at Nature Publishing Group is not mandatory.  This effort
corresponds to an assessment of code availability as suggested by the
Transparency and Openness Promotion (TOP) Guidelines \cite{Nosek2015},
and extends the scope with concrete feedback on core aspects of
long-term reusability.